A lot of fixes

+++ ffn_bot/reddit_bot.py
@@ -47,6 +47,8 @@ FOOTER = "\n\nSupporting fanfiction.net (*linkffn*), AO3 (buggy) (*linkao3*), HP
     "^(**New Feature:** Parse multiple fics in a single call with;semicolons;like;this!)\n\n" + \
     "^^**Update** ^^**7/7/2015:** ^^More ^^formatting ^^bugs ^^fixed.\n\nffnbot!ignore"
 
+DRY_RUN = False
+
 
 def get_regexps():
     """Returns the regular expressions for the sites."""
@@ -68,10 +70,11 @@ def _run_forever():
     while True:
         try:
             main()
+        # Exit on sys.exit and keyboard interrupts.
+        except KeyboardInterrupt:
+            raise
         except SystemExit as e:
             return e.code
-        except KeyboardInterrupt:
-            return 0
         except:
             logging.error("MAIN: AN EXCEPTION HAS OCCURED!")
             bot_tools.print_exception()
@@ -79,6 +82,7 @@ def _run_forever():
 
 
 def main():
+    global DRY_RUN
     """Basic main function."""
     # moved call for agruments to avoid double calling
     bot_parameters = get_bot_parameters()
@@ -86,6 +90,7 @@ def main():
     load_checked_comments()
     load_subreddits(bot_parameters)
 
+    DRY_RUN = bool(bot_parameters["dry"])
     while True:
         for SUBREDDIT in SUBREDDIT_LIST:
             parse_submissions(r.get_subreddit(SUBREDDIT))
@@ -96,24 +101,39 @@ def get_bot_parameters():
     """Parse the command-line arguments."""
     # initialize parser and add options for username and password
     parser = argparse.ArgumentParser()
-    parser.add_argument('-u', '--user', help='define Reddit login username')
-    parser.add_argument('-p', '--password', help='define Reddit login password')
     parser.add_argument(
-        '-s', '--subreddits', help='define target subreddits; seperate with commas')
-
-    # Can also add possibility with -s option to aquire comma separated list of subreddits
-    # then do: subs = args.subreddit.split(',')
-    # and return this list, then append/extend to Subreddit_list or default_subs
-    # and possibl transform to set to avoid duplicates.
-
-    # Alternatively we can just use docopt. (See http://docopt.org/)
+        '-u', '--user',
+        help='define Reddit login username'
+    )
+    parser.add_argument(
+        '-p', '--password',
+        help='define Reddit login password'
+    )
 
     parser.add_argument(
-        '-d', '--default', action='store_true', help='add default subreddits, can be in addition to -s')
+        '-s', '--subreddits',
+        help='define target subreddits; seperate with commas'
+    )
+
+    parser.add_argument(
+        '-d', '--default',
+        action='store_true',
+        help='add default subreddits, can be in addition to -s'
+    )
+
+    parser.add_argument(
+        '-l', '--dry',
+        action='store_true',
+        help="do not send comments."
+    )
 
     args = parser.parse_args()
 
-    return {'user': args.user, 'password': args.password, 'user_subreddits': args.subreddits, 'default': args.default}
+    return {
+        'user': args.user, 'password': args.password,
+        'user_subreddits': args.subreddits, 'default': args.default,
+        'dry': args.dry
+    }
 
 
 def login_to_reddit(bot_parameters):
@@ -144,11 +164,16 @@ def load_subreddits(bot_parameters):
         SUBREDDIT_LIST.add('tusingtestfield')
     print("LOADED SUBREDDITS: ", SUBREDDIT_LIST)
 
-
 def check_comment(id):
     """Marks a comment as checked."""
-    global CHECKED_COMMENTS
+    global CHECKED_COMMENTS, DRY_RUN
     CHECKED_COMMENTS.add(str(id))
+
+    # This is a dry run, do not pretend we
+    # store anything in any way.
+    if DRY_RUN:
+        return
+
     with open('CHECKED_COMMENTS.txt', 'w') as file:
         for id in CHECKED_COMMENTS:
             file.write(str(id) + '\n')
@@ -158,8 +183,12 @@ def load_checked_comments():
     """Loads all comments that have been checked."""
     global CHECKED_COMMENTS
     logging.info('Loading CHECKED_COMMENTS...')
-    with open('CHECKED_COMMENTS.txt', 'r') as file:
-        CHECKED_COMMENTS = {str(line.rstrip('\n')) for line in file}
+    try:
+        with open('CHECKED_COMMENTS.txt', 'r') as file:
+            CHECKED_COMMENTS = {str(line.rstrip('\n')) for line in file}
+    except IOError:
+        bot_tools.print_exception()
+        CHECKED_COMMENTS = set()
     print('Loaded CHECKED_COMMENTS.')
     logging.info(CHECKED_COMMENTS)
 
@@ -200,6 +229,7 @@ def parse_submission_text(submission):
         additions
     )
 
+
 def parse_submissions(SUBREDDIT):
     """Parses all user-submissions."""
     print("==================================================")
@@ -221,7 +251,7 @@ def parse_submissions(SUBREDDIT):
                 print("Parsing comment ", comment.id, ' in submission ', submission.id)
                 try:
                     make_reply(comment.body, comment.id, comment.id, comment.reply)
-                except:
+                except Exception:
                     logging.error("\n\nPARSING COMMENT: Error has occured!")
                     bot_tools.print_exception()
                     check_comment(comment.id)
@@ -241,7 +271,11 @@ def make_reply(body, cid, id, reply_func, markers=None, additions=()):
         print('Outgoing reply to ' + id + ':\n' + reply + FOOTER)
         print('--------------------------------------------------')
         print(Style.RESET_ALL)
-        reply_func(reply + FOOTER)
+
+        # Do not send the comment.
+        if not DRY_RUN:
+            reply_func(reply + FOOTER)
+
         bot_tools.pause(1, 20)
         print('Continuing to parse submissions...')
     else:
